% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance.R
\name{AcceptNewGene}
\alias{AcceptNewGene}
\title{Accepts a new gene.}
\usage{
AcceptNewGene(OperatorPipeline, gene, lF)
}
\arguments{
\item{OperatorPipeline}{Genetic operator pipeline (an R function).}

\item{gene}{Gene.}

\item{lF}{Local configuration.}
}
\value{
New gene.
}
\description{
Executes a genetic operator pipeline.
             The new gene is returned.
}
\examples{
id<-function(x, lF){x}
g1<-InitGene(lFxegaGaGene)
AcceptNewGene(id, g1, lFxegaGaGene)
}
\seealso{
Other Acceptance Rule: 
\code{\link{AcceptBest}()},
\code{\link{AcceptIVMetropolis}()},
\code{\link{AcceptMetropolis}()}
}
\concept{Acceptance Rule}
