% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaEvalPopulation.R
\name{asPipeline}
\alias{asPipeline}
\title{Converts a population into a list of genetic operator pipelines.}
\usage{
asPipeline(pop, lF)
}
\arguments{
\item{pop}{A population.}

\item{lF}{The local function configuration.}
}
\value{
A list of genetic operator pipelines.
}
\description{
Converts a population into a list of genetic operator pipelines.
}
\examples{
pop5<-xegaInitPopulation(5, lFxegaGaGene)
pop5c<-asPipeline(pop5, lFxegaGaGene)
pop5c
}
\seealso{
Other Genetic operator pipelines: 
\code{\link{xegaRepairPop}()}
}
\concept{Genetic operator pipelines}
