% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPipeline.R
\name{asPipelineG}
\alias{asPipelineG}
\title{Embeds genetic operator pipelines into the genes of a population.}
\usage{
asPipelineG(pop, lF)
}
\arguments{
\item{pop}{A population.}

\item{lF}{The local function configuration.}
}
\value{
A population of genes with embedded genetic operator pipelines.
}
\description{
Embeds genetic operator pipelines into the genes of a population.
}
\examples{
pop5<-xegaInitPopulation(5, lFxegaGaGene)
pop5c<-asPipelineG(pop5, lFxegaGaGene)
}
\seealso{
Other Genetic operator pipelines: 
\code{\link{asPipeline}()},
\code{\link{asPipelineID}()},
\code{\link{xegaRepairPop}()}
}
\concept{Genetic operator pipelines}
