% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPipeline.R
\name{asPipelineID}
\alias{asPipelineID}
\title{Identity (No compilation of genetic operator pipelines for population).}
\usage{
asPipelineID(pop, lF)
}
\arguments{
\item{pop}{A population.}

\item{lF}{The local function configuration.}
}
\value{
The population
}
\description{
Identity (No compilation of genetic operator pipelines for population).
}
\examples{
pop5<-xegaInitPopulation(5, lFxegaGaGene)
pop5c<-asPipeline(pop5, lFxegaGaGene)
identical(pop5, pop5c)
}
\seealso{
Other Genetic operator pipelines: 
\code{\link{asPipeline}()},
\code{\link{asPipelineG}()},
\code{\link{xegaRepairPop}()}
}
\concept{Genetic operator pipelines}
