% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate.R
\name{checkTerminatePAC}
\alias{checkTerminatePAC}
\title{Check terminatePAC()}
\usage{
checkTerminatePAC(penv, max)
}
\arguments{
\item{penv}{A problem environment.}

\item{max}{Maximize?}
}
\value{
A named list
        \itemize{
         \item \code{$OK}   \code{TRUE}
         \item \code{$penv} \code{penv}
                }
}
\description{
Check terminatePAC()
}
