% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate.R
\name{checkTerminationFactory}
\alias{checkTerminationFactory}
\title{Configure consistency checks and adapt \code{penv} for terminationConditions.}
\usage{
checkTerminationFactory(method = "NoTermination")
}
\arguments{
\item{method}{A string specifying the termination condition.}
}
\value{
A check function.
}
\description{
For each termination condition, a check must be provided.
             A check fails (stops) if the consistency requirements 
             of a termination condition are not fulfilled.
             However, a check may modify the problem environment to 
             establish consistency.
}
\seealso{
Other Configuration: 
\code{\link{AcceptFactory}()},
\code{\link{ApplyFactory}()},
\code{\link{CoolingFactory}()},
\code{\link{CrossRateFactory}()},
\code{\link{MutationRateFactory}()},
\code{\link{TerminationFactory}()},
\code{\link{xegaAsPipelineFactory}()},
\code{\link{xegaConfiguration}()},
\code{\link{xegaEvalPopulationFactory}()}
}
\concept{Configuration}
