% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptivityCrossover.R
\name{ConstCRate}
\alias{ConstCRate}
\title{Constant crossover rate.}
\usage{
ConstCRate(fit, lF)
}
\arguments{
\item{fit}{Fitness of gene.}

\item{lF}{Local configuration.}
}
\value{
Constant crossover rate.
}
\description{
Constant crossover rate.
}
\examples{
parm<-function(x){function() {return(x)}}
lF<-list(CrossRate1=parm(0.20))
ConstCRate(100, lF)
ConstCRate(50, lF)
}
\seealso{
Other Rates: 
\code{\link{ConstMRate}()}
}
\concept{Rates}
