% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaNextPopulation.R
\name{Cross2Gene}
\alias{Cross2Gene}
\title{Import for examples.}
\usage{
Cross2Gene(gg1, gg2, lF)
}
\arguments{
\item{gg1}{a gene}

\item{gg2}{a gene}

\item{lF}{list of local functions}
}
\value{
a list of two genes
}
\description{
Import for examples.
}
