% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaNextPopulation.R
\name{CrossGene}
\alias{CrossGene}
\title{Import for examples.}
\usage{
CrossGene(gg1, gg2, lF)
}
\arguments{
\item{gg1}{a gene}

\item{gg2}{a gene}

\item{lF}{list of local functions}
}
\value{
a list of one gene
}
\description{
Import for examples.
}
