% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localAdaptivity.R
\name{IAMBitRate}
\alias{IAMBitRate}
\title{Individually adaptive mutation rate. (Bit mutation Rate)}
\usage{
IAMBitRate(fit, lF)
}
\arguments{
\item{fit}{Fitness of gene.}

\item{lF}{Local configuration.}
}
\value{
Mutation rate of a gene depending on its fitness.
}
\description{
Adaptivity of a local operator mutation parameter.
             Currently not used. Implements a threshold rule.
             The rule is implemented directly in IVAdaptiveMutateGene.
             in package xegaGaGene. Move?
}
\details{
TODO: Move this xegaGaGene and generalize the
             bit mutation operator and introduce a factory 
             for bit mutation rates. Rationale: Local parameters
             are representation dependent.
}
