% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaNextPopulation.R
\name{InitGene}
\alias{InitGene}
\title{Import for examples.}
\usage{
InitGene(lF)
}
\arguments{
\item{lF}{a list of local functions}
}
\value{
a new random gene
}
\description{
Import for examples.
}
