% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance.R
\name{MetropolisAcceptanceProbability}
\alias{MetropolisAcceptanceProbability}
\title{Metropolis acceptance probability.}
\usage{
MetropolisAcceptanceProbability(d, beta, temperature)
}
\arguments{
\item{d}{Distance between the fitness of the old and the new gene.}

\item{beta}{Constant.}

\item{temperature}{Temperature.}
}
\value{
Acceptance probability.
}
\description{
Metropolis acceptance probability.
}
\examples{
MetropolisAcceptanceProbability(d=0, beta=1, temperature=10)
MetropolisAcceptanceProbability(d=1, beta=1, temperature=10)
}
\seealso{
Other Diagnostic: 
\code{\link{MetropolisTable}()}
}
\concept{Diagnostic}
