% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance.R
\name{MetropolisTable}
\alias{MetropolisTable}
\title{Metropolis acceptance probability table.}
\usage{
MetropolisTable(d = 1, beta = 2, temperature = 1000, alpha = 0.9, steps = 1000)
}
\arguments{
\item{d}{Distance between the fitness of the old and the new gene.}

\item{beta}{Constant.}

\item{temperature}{Temperature.}

\item{alpha}{Cooling constant in [0, 1].}

\item{steps}{Number of steps.}
}
\value{
Data frame with the columns alpha, beta, temperature, 
        d (distance between fitness), and probability of acceptance.
}
\description{
Metropolis acceptance probability table.
}
\examples{
MetropolisTable(d=2, beta=2, temperature=10, alpha=0.99, steps=10)
}
\seealso{
Other Diagnostic: 
\code{\link{MetropolisAcceptanceProbability}()}
}
\concept{Diagnostic}
