% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parModel.R
\name{PparLapplyHet}
\alias{PparLapplyHet}
\title{uses parLapplyLB of library parallel for using workers on 
machines in a local network.}
\usage{
PparLapplyHet(pop, EvalGene, lF)
}
\arguments{
\item{pop}{a population of genes.}

\item{EvalGene}{the function for evaluating a gene.}

\item{lF}{the local function factory which provides
all functions needed in \code{EvalGene}.}
}
\value{
Fitness vector.
}
\description{
uses parLapplyLB of library parallel for using workers on 
machines in a local network.
}
\section{Warning}{


This section has not been properly tested.
Random number generation?
Examples?
}

\examples{
parm<-function(x) {function() {x}}
pop<-xegaInitPopulation(1000, lFxegaGaGene)
library(parallel)
clus<-makeCluster(2)
lFxegaGaGene$cluster<-parm(clus)
popnew<-PparLapplyHet(pop, lFxegaGaGene$EvalGene, lFxegaGaGene)
stopCluster(clus)

}
\seealso{
Other Execution Model: 
\code{\link{MClapply}()},
\code{\link{MClapplyHet}()},
\code{\link{PparLapply}()},
\code{\link{futureLapply}()},
\code{\link{futureLapplyHet}()}
}
\concept{Execution Model}
