% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaNextPopulation.R
\name{ReplicateGene}
\alias{ReplicateGene}
\title{Import for examples.}
\usage{
ReplicateGene(pop, fit, lF)
}
\arguments{
\item{pop}{the population.}

\item{fit}{the fitness-}

\item{lF}{list of local functions}
}
\value{
a list with one gene
}
\description{
Import for examples.
}
