% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parModel.R
\name{rmpiLapply}
\alias{rmpiLapply}
\title{uses the local function \code{lF$Rmpi$mpi.parLapply()}}
\usage{
rmpiLapply(pop, EvalGene, lF)
}
\arguments{
\item{pop}{a population of genes.}

\item{EvalGene}{the function for evaluating a gene.}

\item{lF}{the local function factory which provides
all functions needed in \code{EvalGene}.}
}
\value{
Fitness vector.
}
\description{
The local function name \code{lF$Rmpi$mpi.parLapply()}
             must be bound to the function mpi.parLapply of package Rmpi
             (Code injection).
}
\seealso{
Other Execution Model: 
\code{\link{MClapply}()},
\code{\link{MClapplyHet}()},
\code{\link{PparLapply}()},
\code{\link{PparLapplyHet}()},
\code{\link{futureLapply}()},
\code{\link{futureLapplyHet}()}
}
\concept{Execution Model}
