% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminate.R
\name{terminatedFalse}
\alias{terminatedFalse}
\title{No termination condition.}
\usage{
terminatedFalse(solution, lF)
}
\arguments{
\item{solution}{A named list with at least the following elements:
$name, $fitness, $value, $numberOfSolutions, 
$genotype, $phenotype, $phenotypeValue.}

\item{lF}{Local function configuration.}
}
\value{
FALSE
}
\description{
A boolean function which always returns \code{FALSE}.
}
\examples{
   lF<-list()
   terminatedFalse(1.0, lF)
}
\seealso{
Other Termination Condition: 
\code{\link{terminateAbsoluteError}()},
\code{\link{terminateGEQ}()},
\code{\link{terminateLEQ}()},
\code{\link{terminatePAC}()},
\code{\link{terminateRelativeError}()},
\code{\link{terminateRelativeErrorZero}()}
}
\concept{Termination Condition}
