% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asPipeline.R
\name{xegaAsPipelineFactory}
\alias{xegaAsPipelineFactory}
\title{Configure asPipeline.}
\usage{
xegaAsPipelineFactory(method = "NoPipe")
}
\arguments{
\item{method}{A string specifying the asPipeline function.}
}
\value{
An asPipeline function.
}
\description{
\code{xegaAsPipelineFactory()} implements the selection
             of one of the asPipeline functions in this
             package by specifying a text string.
             The selection fails ungracefully (produces
             a runtime error) if the label does not match.
             The functions are specified locally.

             Current support:

             \enumerate{
             \item "NoPipe" returns \code{asPipelineID()}:
                   No change to population.
             \item "PipeC" returns \code{asPipeline()}: 
                   Population a list of closures of genetic operator 
                   populations.
             \item "PipeG" returns \code{asPipelineG()}:
                   Population consists of genes with embedded 
                   genetic operator pipelines.
             }
}
\examples{
xegaAsPipelineFactory("PipeC")
}
\seealso{
Other Configuration: 
\code{\link{AcceptFactory}()},
\code{\link{ApplyFactory}()},
\code{\link{CoolingFactory}()},
\code{\link{CrossRateFactory}()},
\code{\link{MutationRateFactory}()},
\code{\link{TerminationFactory}()},
\code{\link{checkTerminationFactory}()},
\code{\link{xegaConfiguration}()},
\code{\link{xegaEvalPopulationFactory}()}
}
\concept{Configuration}
