% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaPopulation.R
\name{xegaInitPopulation}
\alias{xegaInitPopulation}
\title{Initializes a population of genes.}
\usage{
xegaInitPopulation(popsize, lF)
}
\arguments{
\item{popsize}{Population size.}

\item{lF}{Local function configuration.}
}
\value{
List of genes.
}
\description{
\code{xegaInitPopulation()} initializes a population
                of genes.
}
\examples{
pop10<-xegaInitPopulation(10, lFxegaGaGene)

}
\seealso{
Other Population Layer: 
\code{\link{xegaBestGeneInPopulation}()},
\code{\link{xegaBestInPopulation}()},
\code{\link{xegaEvalPopulation}()},
\code{\link{xegaLogEvalsPopulation}()},
\code{\link{xegaNextPopulation}()},
\code{\link{xegaObservePopulation}()},
\code{\link{xegaSummaryPopulation}()}
}
\concept{Population Layer}
