% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaEvalPopulation.R
\name{xegaRepairPop}
\alias{xegaRepairPop}
\title{Repairs the list structure of a population of genes.}
\usage{
xegaRepairPop(pop)
}
\arguments{
\item{pop}{A population of genes.}
}
\value{
A population of genes.
}
\description{
Pipelines with crossover operators with 2 kids 
    generate function closures which return 2 genes (with the 
    complete genetic material). The resulting population has 
    elements with a single gene and elements with a list of 
    two genes.    
    \code{xegaRepairPop} removes this extra nesting structure 
    and returns of population vector of genes of length popsize.
}
\seealso{
Other Genetic operator pipelines: 
\code{\link{asPipeline}()},
\code{\link{asPipelineG}()},
\code{\link{asPipelineID}()}
}
\concept{Genetic operator pipelines}
