% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xegaNextPopulation.R, R/xegaPopulation.R
\docType{data}
\name{lFxegaGaGene}
\alias{lFxegaGaGene}
\title{Import for examples.}
\format{
An object of class \code{list} of length 32.

An object of class \code{list} of length 32.
}
\usage{
lFxegaGaGene

lFxegaGaGene
}
\description{
Import for examples.

Import lFxegaGaGene
}
\keyword{datasets}
